+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Next
 ----
 
 - Add a ``dcos-docker doctor`` check that a supported storage driver is available.
+- Fix error with using Docker version `v17.12.1-ce` inside Docker nodes.
+- Fix race condition between installing DC/OS and SSH starting.
 
 2018.03.07.0
 ------------