+++ CHANGELOG.rst
@@ -6,10 +6,15 @@ Changelog
 Next
 ----
 
+- Add an AWS backend to the library.
+
+2018.03.26.0
+------------
+
 - Add a ``dcos-docker doctor`` check that a supported storage driver is available.
 - Fix error with using Docker version `v17.12.1-ce` inside Docker nodes.
 - Fix race condition between installing DC/OS and SSH starting.
-- Add an AWS backend to the library.
+- Remove support for Ubuntu on the Docker backend.
 
 2018.03.07.0
 ------------